Update callAction() to use less setImmediates
This removes two `setImmediate()` calls in the common action path and
leverages promises more for enforcing async calling of actions.

Since Promises are a standardized EMCAScript feature, JS VMs are
putting effort into improving their performance. While, currently,
`setImmediate()` is a non-standard API.

What we're trying to optimize here is limiting the amount of time the
code is waiting for turns of the JS event loop, while still keeping
the async semantics. The idea is that by using promises the JS engine
will be able to group async operations within the same micro-task
queue, and therefore within the same turn of the event loop.

**Previous Async Turns (Promise-returning Action):**
```
└─┬ setImmediate()
  └─┬ Promise#then()
    └─┬ Promise#then()
      └── setImmediate()
```

**Previous Async Turns (Callback-based Action):**
```
└─┬ setImmediate()
  └─┬ Promise#then()
    └── setImmediate()
```

**New Async Turns (Promise-returning Action):**
```
└─┬ Promise#then()
  └── Promise#then()
```

**New Async Turns (Callback-based Action):**
```
└─┬ Promise#then()
  └── Promise#then()
```