Remove unused historyCreated variable
The 'historyCreated' variable was never set to true, so although it was
intended to ensure that only one history handler was ever created, it
actually had no effect. Setting the variable to true breaks the unit
tests, because they do not all unmount the components used in the test.
Removing the variable altogether preserves current behaviour and avoids
misleading people who read the code.